refactor: disperse cones into state_props / matrix_props / matrix_ops
The cones package mixed three unrelated things under an opaque name: quantum
entropies, generic matrix functions, and CVXPY cone-constraint builders. Move
each to where it conceptually belongs, leaving cones as only the genuine cone
builders.

- state_props (with von_neumann_entropy, quantum_relative_entropy, renyi_entropy):
  ln_quantum_entropy, tsallis_entropy, tsallis_relative_entropy,
  integral_relative_entropy, relative_entropy_quadrature.
- matrix_props (with trace_norm, kp_norm, sk_norm): trace_matrix_log,
  trace_matrix_power, lieb_ando.
- matrix_ops: geometric_mean (returns a matrix). Its is_positive_definite import
  is deferred to keep matrix_ops free of a matrix_props back-edge, preserving the
  documented no-cycle invariant.
- cones now holds only the CVXPY cone-constraint builders (geometric_mean_epi_cone,
  geometric_mean_hypo_cone, operator_relative_entropy_epi_cone) -- so the name is
  finally accurate.

Routing the entropies up into state_props (a top-level package that already
imports cones/matrix_props/matrix_ops) sidesteps the matrix_ops<->matrix_props
cycle entirely, so the only cycle-handling required is the one deferred import in
geometric_mean.

No deprecation shims: the old toqito.cones.<fn> paths are gone. Full suite passes
(3532 passed), import toqito is clean, ruff clean.
